Web/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework>=3.14.0
Flask==2.2.2
Flask-RESTful==0.3.9
Pillow==9.3.0
psycopg2-binary==2.9.5
python-decouple==3.6
requests==2.26
sqlparse==0.4.3
tzdata==2022.6
